// eltvpp-ui/src/store/modules/app/types.ts
import type { RouteRecordNormalized } from "vue-router";

export interface AppState {
  theme: string;
  colorWeak: boolean;
  navbar: boolean;
  menu: boolean;
  topMenu: boolean;
  hideMenu: boolean;
  menuCollapse: boolean;
  footer: boolean;
  themeColor: string;
  menuWidth: number;
  globalSettings: boolean;
  device: string;
  tabBar: boolean;
  menuFromServer: boolean;
  serverMenu: RouteRecordNormalized[];
  serverMenuAll: RouteRecordNormalized[];
  fatherMenu: any[];
  lastSign: string;
  lastIndex: number;
  [key: string]: unknown;
}